backend/requirements.txt
# Django
Django==5.0.1
djangorestframework==3.14.0
django-filter==23.5

# Database
psycopg2-binary==2.9.9

# Authentication
djangorestframework-simplejwt==5.3.1
pyotp==2.9.0
qrcode==7.4.2

# Encryption & Security (for multi-org consent)
cryptography==42.0.0
bcrypt==4.1.2

# CORS
django-cors-headers==4.3.1

# Async/WebSocket
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Celery
celery==5.3.4
redis==5.0.1

# File Storage
boto3==1.34.22
django-storages==1.14.2

# PDF Generation
WeasyPrint==60.2
jinja2==3.1.3

# Excel Export
openpyxl==3.1.2

# API Documentation
drf-spectacular==0.27.0

# Code Quality
black==23.12.1
isort==5.13.2
ruff==0.1.11

# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==22.0.0

# Utilities
python-dotenv==1.0.0
pillow==10.2.0
requests==2.31.0

# AI Integration
google-generativeai==0.3.2

